Rewrote persistent cache;
Added DecodeManager logging;
Improved thread cancelling so results from canceled thread is never sent back anymore;
Using C++20;
Fixed bug where cached results were overwriting decoded ones;
Image info from cache is restored in bulk and so is applied to MasonryLayout;
Fixed viewer decoding: it wouldn't request the same thing over and over again;
Not decoding or laying out anything until window is fully initialized and shown;
Fixed crash when vigorously refreshing view;
Images are now decoded to fit cached resolution and then scaled to target resolution;
Supported orientation for Qt decoder;
Fixed broken image selection in right side panel;

+++ exiv2/Preview.cpp
@@ -1433,13 +1433,13 @@ Exiv2::DataBuf Exiv2Preview::preview(const Exiv2::Image& image, int targetWidth,
                 continue;
             }
             PreviewProperties props = loader->getProperties();
-            if (props.width_ > maxPreviewWidth && props.height_ > maxPreviewHeight) {
+            if ((int)props.width_ > maxPreviewWidth && (int)props.height_ > maxPreviewHeight) {
                 maxPreviewWidth = props.width_;
                 maxPreviewHeight = props.height_;
                 previewWithMaxResolution = id;
             }
 
-            if (props.width_ > targetWidth && props.height_ > targetHeight) {
+            if ((int)props.width_ > targetWidth && (int)props.height_ > targetHeight) {
                 properties = loader->getProperties();
                 buf = loader->getData(); // #16 getPreviewImage()
                 properties.size_ = buf.size_;         //     update the size
